DUPLICATE CREATOR now appends only to notes columns corresponding to videos uploaded by the duplicate creator.
